* Left half letters working, shift key working (capital letters) * vial.json using layout from www.keyboard-layout-editor.com * Right half of keyboard now working for letters and also correct in Vial
Pausing development in QMK having found that, because keyboard consists
of two independent halves, the Shift/Alt/Ctrl keys one half won't be
registered against keystrokes made on the other. Also layer changes
won't take effect, etc.

In view of this I have decided to switch to a wireless keyboard using
nRF52840 microcontrollers, they have the same footprint with compatible
pinouts as the Frood microcontrollers and so I should be able to use
them as substitutes. Switching to wireless means the two halves will be
able to communicate/coordinate and so the keyboard will be registered as
a single device and so the modifiers, layer-change keys, etc. will
function as intended. However it also means I will need to find
alternative firmware to QMK/Vial as QMK doesn't support the nRF52840
microcontroller.